fal-inference/requirements.txt
fal-serverless>=0.7
torch>=2.2.0
diffusers>=0.28.0
transformers>=4.40.0
accelerate>=0.29.0
pydantic>=2.0
imageio[ffmpeg]>=2.34.0
Pillow>=10.0
huggingface-hub>=0.22.0
